voice-agents/pyproject.toml
[project]
name = "voice-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.2.3",
    "livekit-plugins-bithuman>=1.2.3",
    "livekit-plugins-google>=1.2.3",
    "livekit-plugins-silero>=1.2.3",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]
